Move resolve submobules into their parents
