+++ betty/rich/user.py
@@ -15,7 +15,7 @@ from typing_extensions import override
 
 from betty.assertion import Assertion
 from betty.locale.localizable import ResolvableLocalizable
-from betty.locale.localize.resolve import resolve_localized
+from betty.locale.localize import resolve_localized
 from betty.progress import Progress
 from betty.progress.no_op import NoOpProgress
 from betty.rich import Theme
