+++ betty/test_utils/user.py
@@ -12,8 +12,7 @@ from typing_extensions import override
 
 from betty.assertion import Assertion
 from betty.locale.localizable import ResolvableLocalizable
-from betty.locale.localize import DEFAULT_LOCALIZER
-from betty.locale.localize.resolve import resolve_localized
+from betty.locale.localize import DEFAULT_LOCALIZER, resolve_localized
 from betty.progress import Progress
 from betty.progress.no_op import NoOpProgress
 from betty.typing import Void, internal
