+++ betty/tests/locale/localize/test___init__.py
@@ -1,16 +1,29 @@
 from __future__ import annotations
 
+from gettext import NullTranslations
 from typing import TYPE_CHECKING
 
 from babel import Locale
 
-from betty.locale.localize import DEFAULT_LOCALIZER, LocalizerRepository
+from betty.locale import DEFAULT_LOCALE, DEFAULT_LOCALE_TAG
+from betty.locale.localizable.plain import Plain
+from betty.locale.localize import (
+    DEFAULT_LOCALIZER,
+    Localizer,
+    LocalizerRepository,
+    resolve_localized,
+)
 from betty.locale.translation import TranslationRepository
 
 if TYPE_CHECKING:
     from pathlib import Path
 
     from pytest_mock import MockerFixture
+
+    from betty.locale.localizable import (
+        ShorthandStaticTranslations,
+        StaticTranslationsMapping,
+    )
 
 
 class TestLocalizer:
@@ -87,3 +100,38 @@ class TestLocalizerRepository:
         localizer = sut.get(locale)
         assert localizer.locale == Locale(locale)
         assert sut.get(locale) is localizer
+
+
+def test_resolve_localized__with_localizable() -> None:
+    localizable = "My First Localizable"
+    assert (
+        resolve_localized(Plain(localizable), localizer=DEFAULT_LOCALIZER)
+        == localizable
+    )
+
+
+def test_resolve_localized__with_str() -> None:
+    localizable = "My First Localizable"
+    assert resolve_localized(localizable, localizer=DEFAULT_LOCALIZER) == localizable
+
+
+def test_resolve_localized__with_static_translations_mapping() -> None:
+    locale = "nl"
+    localizer = Localizer(locale, NullTranslations())
+    localized = "Mijn Eerste, Ja, Wat Eigenlijk?"
+    localizable: StaticTranslationsMapping = {
+        DEFAULT_LOCALE: "My First Localizable",
+        Locale(locale): localized,
+    }
+    assert resolve_localized(localizable, localizer=localizer) == localized
+
+
+def test_resolve_localized__with_shorthand_static_translations_mapping() -> None:
+    locale = "nl-NL"
+    localizer = Localizer(locale, NullTranslations())
+    localized = "Mijn Eerste, Ja, Wat Eigenlijk?"
+    localizable: ShorthandStaticTranslations = {
+        DEFAULT_LOCALE_TAG: "My First Localizable",
+        locale: localized,
+    }
+    assert resolve_localized(localizable, localizer=localizer) == localized
